Renamed Decorator to Formatter.
