+++ sdoc/sdoc2/formatter/html/HeadingHtmlFormatter.py
@@ -8,7 +8,7 @@ Licence MIT
 # ----------------------------------------------------------------------------------------------------------------------
 from sdoc.helper.Html import Html
 from sdoc.sdoc2 import node_store
-from sdoc.sdoc2.decorator.html.HtmlFormatter import HtmlFormatter
+from sdoc.sdoc2.formatter.html.HtmlFormatter import HtmlFormatter
 
 
 class HeadingHtmlFormatter(HtmlFormatter):
